requirements.txt
llama-index
llama-index-embeddings-nebius
llama-index-vector-stores-chroma  
pypdf
sentence-transformers
chromadb        
python-dotenv
yfinance
llama-index-llms-nebius
fastapi
uvicorn[standard]